Implement updating for GitHub projects & move to its own action
